refactor: redo Tree internals
BREAKING CHANGE: remove Deref/DerefMut impl to removed TreeInterface
